+++ src/test/java/tk/mybatis/mapper/mapper/MybatisHelper.java
@@ -75,6 +75,8 @@ public class MybatisHelper {
                 config.setNotEmpty(true);
                 //校验Example中的类型是否一致
                 config.setCheckExampleEntityClass(true);
+                //启用简单类型
+                config.setUseSimpleType(true);
                 // 序列的获取规则,使用{num}格式化参数，默认值为{0}.nextval，针对Oracle
                 // 可选参数一共3个，对应0,1,2,分别为SequenceName，ColumnName, PropertyName
                 //config.setSeqFormat("NEXT VALUE FOR {0}");