+++ wiki/Changelog.md
@@ -3,8 +3,10 @@
 ##3.4.0-SNAPSHOT - 2017-01-xx
 
 * `Example` 增加 for update 支持，仅能用于 selectByExample 和 selectCountByExample 方法 #210
-* 增加参数`checkExampleEntityClass`，用于校验通用 Example 构造参数 entityClass 是否和当前调用的 Mapper<EntityClass> 类型一致 #201
 * `Example.Criteria` 增加 `andAllEqualTo` 方法，将此对象的所有字段参数作为相等查询条件，如果字段为 null，则为 is null #206
+* 增加参数 `checkExampleEntityClass`，默认 `false` 用于校验通用 Example 构造参数 entityClass 是否和当前调用的 Mapper<EntityClass> 类型一致 #201
+* 增加参数 `useSimpleType`，默认 `false`，启用后判断实体类属性是否为表字段时校验字段是否为简单类型，如果不是就忽略该属性，这个配置优先级高于所有注解
+* 增加参数 `simpleTypes`，默认的简单类型在 `SimpleTypeUtil` 中，使用该参数可以增加额外的简单类型，通过逗号隔开的全限定类名添加
 
 ##3.3.9 - 2016-09-04